+++ src/App.css
@@ -1,13 +1,5 @@
-#root {
-  width :100vw;
-}
-
 body {
   margin: 0;
   height: 100vh;
   width: 100%;
 }
-/* 
-.App {
-  height: 100vh;
-} */
