+++ src/layouts/Content.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Content.css";
+
 
 function Content() {
   return <div className="content">Content</div>;
