--- src/layouts/content/Content.css
@@ -1,3 +1,0 @@
-/* .content {
-  height: 70vh;
-} */
