--- src/layouts/footer/Footer.css
@@ -1,14 +1,0 @@
-.footer {
-  position: fixed;
-	bottom: 0;
-  left: 0;
-  background-color: #343B41;
-  color: #F7F7F7;
-  text-align: center;
-  width: 100%;
-}
-
-.footer p {
-  display: inline-block;
-  padding: 0 5%;
-}
