--- src/layouts/footer/Footer.js
@@ -1,13 +1,0 @@
-import React from "react";
-import "./Footer.css";
-
-function Footer() {
-  return (
-    <div className="footer">
-      <p>Copyright 2021</p>
-      <p>Ochrona danych osobowych</p>
-      <p>Polityka prywatności</p>
-    </div>);
-}
-
-export default Footer;
