--- src/layouts/header/Header.css
@@ -1,10 +1,0 @@
-.header {
-  /* position: absolute; */
-  width: 100%;
-  height: 20vh;
-}
-
-.header p {
-    display: inline-block;
-    padding: 0 5%;
-  }
