--- src/layouts/layout/Layout.css
@@ -1,3 +1,0 @@
-/* .layout-container {
-  height: 100%;
-} */
