--- src/layouts/logo/Logo.css
@@ -1,3 +1,0 @@
-.logo {
-  height: 60%;
-}
