--- src/layouts/logo/Logo.js
@@ -1,8 +1,0 @@
-import React from "react";
-import "./Logo.css";
-
-function Logo() {
-  return <div className="logo">Logo</div>;
-}
-
-export default Logo;
