--- src/layouts/navbar/Navbar.css
@@ -1,13 +1,0 @@
-.navbar {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: #343B41;
-  color: #F7F7F7;
-  height: 40%;
-}
-
-/* .navbar p {
-  display: inline-block;
-  padding: 0 5%;
-} */
