--- src/layouts/navbar/Navbar.js
@@ -1,17 +1,0 @@
-import React from "react";
-import "./Navbar.css";
-
-function Navbar() {
-  return (
-    <div className="navbar">
-      <p>O Nas</p>
-      <p>Usługi</p>
-      <p>Nasz zespół</p>
-      <p>Umów wizytę</p>
-      <p>FAQ</p>
-      <p>Kontakt</p>
-    </div>
-    );
-}
-
-export default Navbar;
